+++ resolwe_bio/kb/serializers.py
@@ -8,7 +8,6 @@ Serializers
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from rest_framework import serializers
-from drf_haystack.serializers import HaystackSerializerMixin
 
 from .models import Feature, Mapping
 
@@ -23,18 +22,6 @@ class FeatureSerializer(serializers.ModelSerializer):
         fields = '__all__'
 
 
-class FeatureAutocompleteSerializer(HaystackSerializerMixin, FeatureSerializer):
-    """Feature autocomplete serializer."""
-
-    class Meta(FeatureSerializer.Meta):
-        """Meta configuration for the feature autocomplete serializer."""
-
-        search_fields = ('genes_auto',)
-        field_aliases = {
-            'query': 'genes_auto',
-        }
-
-
 class MappingSerializer(serializers.ModelSerializer):
     """Serializer for mapping."""
 
@@ -43,12 +30,3 @@ class MappingSerializer(serializers.ModelSerializer):
 
         model = Mapping
         fields = '__all__'
-
-
-class MappingSearchSerializer(HaystackSerializerMixin, MappingSerializer):
-    """Mapping search serializer."""
-
-    class Meta(MappingSerializer.Meta):
-        """Meta configuration for the mapping search serializer."""
-
-        search_fields = ('relation_type', 'source_db', 'source_id', 'target_db', 'target_id')