+++ docs/CHANGELOG.rst
@@ -29,6 +29,7 @@ Changed
 * Add source query parameter to kb autocomplete
 * GO enrichment analysis: map features using gene Knowledge base
 * Add option to upload .gff v2 files with upload-gtf processor
+* Replace Haystack with Resolwe Elastic Search API
 
 Added
 -----